add description. delete controllers auth. delete cookieParser
