+++ app.js
@@ -2,13 +2,11 @@ require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
-const cookieParser = require('cookie-parser');
 const auth = require('./middlewares/auth');
 
 const { PORT = 3000, BASE_PATH } = process.env;
 
 const app = express();
-app.use(cookieParser());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
