--- controllers/auth.js
@@ -1,51 +1,0 @@
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const User = require('../models/user');
-
-const { NODE_ENV, JWT_SECRET } = process.env;
-
-function login(req, res) {
-  const { email, password } = req.body;
-
-  return User.findUserByCredentials(email, password)
-    .then((user) => {
-      const token = jwt.sign(
-        { _id: user._id },
-        NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret',
-        { expiresIn: '7d' },
-      );
-      res.cookie('jwt', token, {
-        maxAge: 3600000,
-        httpOnly: true,
-        sameSite: true,
-      })
-        .end();
-    })
-    .catch((err) => {
-      res.status(401).send({ message: err.message });
-    });
-}
-
-function createUser(req, res) {
-  bcrypt.hash(req.body.password, 10)
-    .then((hash) => User.create({
-      name: req.body.name,
-      about: req.body.about,
-      avatar: req.body.avatar,
-      email: req.body.email,
-      password: hash,
-    }))
-    .then((user) => {
-      res.status(201).send({
-        _id: user._id,
-        email: user.email,
-      });
-    })
-    .catch((err) => {
-      res.status(400).send(err);
-    });
-}
-
-module.exports = {
-  login, createUser,
-};
