+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "lesson-twelve",
   "version": "1.0.0",
-  "description": "",
+  "description": "interactive page where you can add photos, delete them and like",
   "main": "index.js",
   "scripts": {
     "dev": "nodemon app.js",
