+++ routes/auth.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { createUser, login } = require('../controllers/auth');
+const { createUser, login } = require('../controllers/users');
 
 router.post('/signin', login);
 router.post('/signup', createUser);
